Adds save file, AI player, & preparation of advanced play
Saves/Reads the ArrayList which contains NimPlayer objects directly into/from a file of which name is "players.dat".

AI players are instances of class NimAIPlayer.
In their rounds of a normal Nim game, AI players will remove the numbers of stones which offers the best chances for them to win.

Advanced play is not added yet.